faster "medium" players, slower "huge" players
Previous patch modified largeModifier to increase it. It boosted
attacker loss but also increased speed a lot (not intentional).
We now declare two different modifier for speed and for attacker loss.
Attacker loss should remain the same.
Speed was brought as before (to slow down a bit huge players). Except
instead of sqrt (exponent 1/2) we do exponent 0.6. So the speed modifier
decreases a bit faster with size (to disadvantage a bit huge players who
are invincible right now)