SAM rebalance (circle radius, faster speed)
First missile are too slow sometimes which is weird when they don't
catch up with missile so I increased speed. Made the range a circle.
75 circle radius is more than 100 for square.
example (60,60) (10, 10)
(60-10)+(60-10)=100 (old formula with the range of 100)
(60-10)^2+(60-10)^2<75^2 (new formula with the radius of 75)

Instead of 10% hit made it a 50% hit. The reasoning is as follow: sam
counter is forcing a cooldown (7.5s) by launching a nuke. If player
fails to counter sam with a nuke then it's fair it's H bomb has 50%
chance of dying. If it's only 10% then SAM is totally useless as almost
any H bomb will go through.