Calculate 'number of teams' and 'max players' at same time
Previous we calculated the number of teams,
then separately calculated the maximum players (defined for a Team game
as 1.5x the recommended map amount).

This resulted in uneven teams, e.g. 50 players split across 6 teams
would leave 4 teams with 8 players and 2 teams with 9 players,
which is unfair.

So now we calculate the number of players based on the number of teams,
ensuring that the number of players results in even teams.

FFA is unchanged.
Lobbies are unchanged -- they can have unbalanced teams if they want.