New Unit Test, Update to Sanitization Funtion
New unit test to confirm that the censor is working as intended. The test will mock the behavior of the obscenity library so that we can arbitrarily set "bad" words for testing.

It also mocks the translation function to work around document.getElement errors are avoided.

The sanitization logic was dropping characters such as "Ü" rather than replacing them with plaintext.