Ensuring that MutableAlliance::extend is called in test case
A test case for AllianceExtensionExecution was not actually checking if the alliance was extended correctly. The method to extend the alliance, MutableAlliance::extend, was never called during execution of AllianceExtensionExecution. The reason the extend() method was not being called is because both mock players were not alive. This PR spies on the MutableAlliance object to verify that extend() is correctly called. Related: https://github.com/openfrontio/OpenFrontIO/pull/1536