fix pattern modal not opening on click (#1826)
## Description:

Sometimes clicking the pattern button wouldn't open the pattern modal.
This is because there was a race condition between opening the modal &
the o-modal becoming available.

So we do `await this.updateComplete;` to ensure the o-modal is
initialized before rendering.
## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [ ] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced

## Please put your Discord username so you can be contacted if a bug or
regression is found:

evan