FIX/IMPROVE Clan Identification
Removed the "^" from the clan identification to ensure [?????] is identified anywhere in the username. Also added 0-9 in the character detect to allow integers in the clan name. 

This is consistent with the above logic: clan = null is assigned when !name.startsWith("[") || !name.contains("]"). Therefore, it stands to reason from the second part of that if statement that if `]` is found within the player name, it *could* be identified as a clan tag. 

This change to Clan identification matching will improve that functionality.