Fix: Resolve deployment failure due to euclideanDist
Corrected the usage of distance calculation in MIRVExecution.ts.
Previously, euclideanDist was called on the Game object, which does not exist.
Changed to use Math.sqrt(this.mg.euclideanDistSquared()) which is the correct method.

This resolves the TS2339: Property 'euclideanDist' does not exist on type 'Game'. error that caused the deployment to fail.