throttle renderGame() based on the current backlog (again)
Rendering runs on its own rAF loop and always tries to hit the display framerate, so even though it’s on a different thread it can still chew a lot of CPU while the worker is trying to catch up a huge backlog. To mitigate this, we now emit backlog status from ClientGameRunner and throttle renderGame() based on the current backlog, effectively reintroducing frame skipping so the worker and main-thread update processing get more breathing room when we’re far behind.